+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-06-14
+
+### Added
+
+- Added curated bundled note fonts: Kavoon, Londrina Shadow, Nabla, Press Start
+  2P and Style Script.
+- Added a Monospace note font option.
+
+### Changed
+
+- Reordered the note font menu to show default/system fonts first, separated
+  from bundled custom fonts.
+- Changed font menu options to render each font label using its corresponding
+  font.
+
 ## [0.2.1] - 2026-06-14
 
 ### Added
@@ -66,7 +81,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added GitHub Pages project page with screenshots.
 - Added repository contributor guidelines.
 
-[Unreleased]: https://github.com/robmoraes/pinleaf/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/robmoraes/pinleaf/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/robmoraes/pinleaf/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/robmoraes/pinleaf/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/robmoraes/pinleaf/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/robmoraes/pinleaf/releases/tag/v0.1.0